% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R
\docType{data}
\name{castlemaine}
\alias{castlemaine}
\title{castlemaine}
\format{List of \code{data.frame} objects representing different components of the
street network}
\description{
'Silicate' ('SC') format street network data for Castlemaine, Australia.
}
\keyword{datasets}
