% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R
\docType{data}
\name{castlemaine_attr}
\alias{castlemaine_attr}
\title{castlemaine_attr}
\format{A \code{data.frame} of point locations of activity attractors}
\description{
Centres of activity attraction in Castlemaing, Australia.
}
\keyword{datasets}
