% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R
\docType{data}
\name{castlemaine_green}
\alias{castlemaine_green}
\title{castlemaine_green}
\format{\pkg{sf}-format \code{data.frame} of green space polygons}
\description{
Polygons of all green spaces in Castlemaine
}
\keyword{datasets}
