% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green-space.R
\name{get_green_space}
\alias{get_green_space}
\title{get_green_space}
\usage{
get_green_space(city, quiet = FALSE)
}
\arguments{
\item{city}{City for which green space polygons are to be extracted.}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\value{
An \pkg{sf}-format \code{data.frame} of polygons representing all green
areas.
}
\description{
Get polygons of all green areas for a given location
}
