% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{move_distances}
\alias{move_distances}
\title{move_distances}
\usage{
move_distances(graph, from, to, wt_profile = "bicycle", expand = 0,
  parallel = TRUE, quiet = TRUE)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph (see Details)}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Details)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Details)}

\item{wt_profile}{Name of weighting profile for street networks (one of foot,
horse, wheelchair, bicycle, moped, motorcycle, motorcar, goods, hgv, psv).}

\item{expand}{Only when \code{graph} not given, the multiplicative factor by
which to expand the street network surrounding the points defined by
\code{from} and/or \code{to}.}

\item{parallel}{If \code{TRUE}, perform routing calculation in parallel (see
details)}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
square matrix of distances between nodes
}
\description{
Alias for \link{move_dists}
}
\examples{
# A larger example from the included [hampi()] data.
graph <- dodgr::weight_streetnet (dodgr::hampi)
from <- sample (graph$from_id, size = 100)
to <- sample (graph$to_id, size = 50)
d <- move_dists (graph, from = from, to = to)
# d is a 100-by-50 matrix of distances between `from` and `to`
}
