% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-stats.R
\name{move_stats}
\alias{move_stats}
\title{move_stats}
\usage{
move_stats(graph, from, d_threshold = 1, quiet = TRUE)
}
\arguments{
\item{graph}{An \code{dodgr_streetnet} object}

\item{from}{Vector of points from which moveability statistics are to be be
calculated.}

\item{d_threshold}{Distance threshold below which moveability statistics are
to be aggreagted (in kilometres).}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
Vector of moveability values for each point in \code{from}, with
moveability quantified as \code{$m}.
}
\description{
Calculate vector of moveability statistics for a given input street network.
}
\examples{
# A larger example from the included [hampi()] data.
graph <- dodgr::weight_streetnet (dodgr::hampi)
from <- sample (graph$from_id, size = 100)
d <- move_stats (graph, from = from)
# d is a `data.frame` of the coordinates of all `from` points and
# correponding moveability statisics 
}
