% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R, R/moveability.R
\docType{package}
\name{moveability}
\alias{moveability}
\alias{moveability-package}
\title{moveability.}
\usage{
moveability(streetnet = NULL, green = NULL, d_threshold = 1,
  mode = "foot", quiet = FALSE)
}
\arguments{
\item{streetnet}{Instead of city, a pre-downloaded or prepared street network
can be submitted. Must be either an \pkg{sf}, \pkg{osmdata} or \pkg{dodgr}
format.}

\item{green}{Areas of green space obtained from \link{get_green_space}}

\item{d_threshold}{Distance threshold below which distances are to be
aggreagted (in kilometres).}

\item{mode}{Mode of transport: either "foot" or "bicycle"}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\value{
Nothing (open interactive map)
}
\description{
Moveability statistics for any specified city.

Calculate moveability statistics for a specified city
}
\examples{
m <- moveability (streetnet = castlemaine)
}
