% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-formats.R
\name{moveability_to_lines}
\alias{moveability_to_lines}
\title{moveability_to_lines}
\usage{
moveability_to_lines(m, streetnet)
}
\arguments{
\item{m}{Result of \link{moveability} function}

\item{streetnet}{Street network in either \pkg{sf}, \pkg{osmdata} or \pkg{dodgr}
format.}
}
\value{
\pkg{sf} collection of \code{LINE} objects of the street network,
with moveability statistics averaged between the two end points of each line
segment.
}
\description{
Project moveability statistics from \link{moveability} on to lines
of street network.
}
\examples{
m <- moveability (streetnet = castlemaine, green_polys = castlemaine_green)
l <- moveability_to_lines (m = m, streetnet = castlemaine)
}
