% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-formats.R
\name{moveability_to_polygons}
\alias{moveability_to_polygons}
\title{moveability_to_polygons}
\usage{
moveability_to_polygons(m, streetnet)
}
\arguments{
\item{m}{Result of \link{moveability} function}

\item{streetnet}{Street network in either \pkg{sf}, \pkg{osmdata} or \pkg{dodgr}
format.}
}
\value{
\pkg{sf} collection of \code{POLYGON} objects corresponding to street
blocks, with average moveability statistics from all points defining that
block.
}
\description{
Project moveability statistics from \link{moveability} on to polygonal blocks
of street network, in order to plot polygons rather than points.
}
\note{
This function may take a long time to execute, because of the
calculation of fundamental cycles in the street network graph used to
identify street blocks.
}
\examples{
m <- moveability (streetnet = castlemaine, green_polys = castlemaine_green)
p <- moveability_to_polygons (m = m, streetnet = castlemaine)
}
