% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability.R
\name{polygons_to_sf}
\alias{polygons_to_sf}
\title{polygons_to_sf}
\usage{
polygons_to_sf(polygons)
}
\arguments{
\item{polygons}{Result of \link{moveability_to_polygons} function}
}
\value{
Equivalent \pkg{sf} collection of \code{POLYGON} objects corresponding to
street blocks, with average moveability statistics from all points defining
that block.
}
\description{
Convert polygons produced from \link{moveability_to_polygons} into equivalent
\pkg{sf} format.
}
\examples{
m <- moveability (streetnet = dodgr::hampi)
p <- moveability_to_polygons (m = m, streetnet = dodgr::hampi)
psf <- polygons_to_sf (p)
mvals <- unlist (lapply (p, function (i) i$m [1]))
psf <- sf::st_sf (dat = mvals, geometry = psf)
}
