% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal-stats.R
\name{get_attractors}
\alias{get_attractors}
\title{get_attractors}
\usage{
get_attractors(bbox, quiet = FALSE)
}
\arguments{
\item{bbox}{Bounding box for which green space polygons are to be extracted.}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\description{
Get points of trip attraction
}
