% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{move_dists}
\alias{move_dists}
\title{movedists}
\usage{
move_dists(graph, from, quiet = TRUE)
}
\arguments{
\item{graph}{An \code{dodgr_streetnet} object}

\item{from}{Vector of points from which moveability statistics are to be be
calculated.}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
Vector of moveability values for each point in \code{from}
}
\description{
Calculate matrix of pair-wise distances between points.
}
\examples{
# A larger example from the included [hampi()] data.
graph <- dodgr::weight_streetnet (dodgr::hampi)
from <- sample (graph$from_id, size = 100)
d <- move_dists (graph, from = from)
# d is a 100-by-50 matrix of distances between `from` and `to`
}
