% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-stats.R
\name{move_statistics}
\alias{move_statistics}
\title{move_statistics}
\usage{
move_statistics(graph, from, green_polys, d_threshold = 1,
  quiet = TRUE)
}
\arguments{
\item{graph}{Street network in \pkg{dodgr} format obtained through applying
\code{dodgr::weight_streetnet} to an \code{osmdata_sc} object.}

\item{from}{Vector of points from which moveability statistics are to be be
calculated.}

\item{green_polys}{Polygons of green space obtained from
\link{get_green_space}}

\item{d_threshold}{Distance threshold below which distances are to be
aggreagted (in kilometres).}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\value{
Vector of moveability values for each point in \code{from}, with
moveability quantified as \code{$m}.
}
\description{
Alias for \link{move_stats}
}
\examples{
graph <- dodgr::weight_streetnet (castlemaine)
green_polys <- castlemaine_green # green polygon data included with package
activity_points <- castlemaine_attr # activity attractors included with package
from <- sample (graph$.vx0, size = 100)
d <- move_stats (graph, green_polys = green_polys,
                 activity_points = activity_points, from = from)
# d is a `data.frame` of the coordinates of all `from` points and
# correponding moveability statisics 
}
