% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R, R/moveability.R
\docType{package}
\name{moveability}
\alias{moveability}
\alias{moveability-package}
\title{moveability.}
\usage{
moveability(streetnet = NULL, from = NULL, green_polys = NULL,
  activity_points, d_threshold = 1, mode = "foot", quiet = FALSE)
}
\arguments{
\item{streetnet}{Pre-downloaded or prepared street network in either
\code{osmdata_sc} or \code{dodgr_sc} format.}

\item{from}{If provided, calculate moveability statistics only at these
points (given as a vector of OSM IDs).}

\item{green_polys}{Polygons of green space obtained from
\link{get_green_space}}

\item{activity_points}{Points of activity obtained from
\link{get_attractors}.}

\item{d_threshold}{Distance threshold below which distances are to be
aggreagted (in kilometres).}

\item{mode}{Mode of transport: either "foot" or "bicycle"}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\value{
Nothing (open interactive map)
}
\description{
Moveability statistics for any specified city.

Calculate moveability statistics for a specified city
}
\examples{
m <- moveability (streetnet = castlemaine, green_polys = castlemaine_green,
                  activity_points = castlemaine_attr)
}
