% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveability-package.R, R/moveability.R
\docType{package}
\name{moveability}
\alias{moveability}
\alias{moveability-package}
\title{moveability.}
\usage{
moveability(city = NULL, streetnet = NULL, d_threshold = 1,
  quiet = FALSE)
}
\arguments{
\item{city}{City for which moveability statistics are to be calcualted.}

\item{streetnet}{Instead of city, a pre-downloaded or prepared street network
can be submitted. Must be either an \pkg{sf}, \pkg{osmdata} or \pkg{dodgr}
format.}

\item{d_threshold}{Distance threshold below which distances are to be
aggreagted (in kilometres).}

\item{quiet}{If \code{TRUE}, dump progress information to screen.}
}
\value{
Nothing (open interactive map)
}
\description{
Moveability statistics for any specified city.

Calculate moveability statistics for a specified city
}
